# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * LastMinute Tours API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TourSearchRequest {
  /** City of departure (e.g. "Москва") */
  departureCity: string;
  /** Maximum budget per person in rubles */
  budget: number;
  /**
   * Number of adults (1-10)
   * @minimum 1
   * @maximum 10
   */
  adults?: number;
}

export interface Tour {
  id: string;
  /** Destination country and city */
  destination: string;
  country: string;
  city: string;
  hotel: string;
  stars: number;
  /** ISO date string */
  departureDate: string;
  /** ISO date string */
  returnDate: string;
  nights: number;
  /** Price per person in rubles */
  price: number;
  /** Total price for all adults in rubles */
  totalPrice: number;
  /** Meal type (e.g. "Всё включено") */
  mealType: string;
  imageUrl?: string;
  /** AI-generated description of what to see and whether to book */
  aiDescription: string;
  /** Short AI verdict - стоит или нет */
  aiRecommendation: string;
  /** URL to booking page on level.travel or partner site */
  bookingUrl: string;
}

export interface TourSearchResponse {
  tours: Tour[];
  searchId?: string;
  departureCity: string;
  budget: number;
  totalFound: number;
}

export interface City {
  id: string;
  name: string;
  nameEn: string;
}

export interface CitiesResponse {
  cities: City[];
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

/**
 * Opaque session token — Bearer <sid>.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
  returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
  code?: string;
  state?: string;
};