# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LastMinute Tours API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: tours
    description: Tour search operations
  - name: Auth
    description: Browser and mobile authentication endpoints.
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /tours/search:
    post:
      operationId: searchTours
      tags: [tours]
      summary: Search last-minute tours
      description: Search for tours departing within 7 days from a given city within a budget
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TourSearchRequest"
      responses:
        "200":
          description: Top 3 tour results with AI descriptions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TourSearchResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /tours/cities:
    get:
      operationId: getDepartureCities
      tags: [tours]
      summary: Get available departure cities
      description: Returns list of supported departure cities
      responses:
        "200":
          description: List of departure cities
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CitiesResponse"
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the returnTo path on success.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '302':
          description: Redirect to the OIDC provider logout URL.
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — Bearer <sid>.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    TourSearchRequest:
      type: object
      properties:
        departureCity:
          type: string
          description: City of departure (e.g. "Москва")
        budget:
          type: number
          description: Maximum budget per person in rubles
        adults:
          type: integer
          description: Number of adults (1-10)
          default: 2
          minimum: 1
          maximum: 10
      required:
        - departureCity
        - budget
    Tour:
      type: object
      properties:
        id:
          type: string
        destination:
          type: string
          description: Destination country and city
        country:
          type: string
        city:
          type: string
        hotel:
          type: string
        stars:
          type: integer
        departureDate:
          type: string
          description: ISO date string
        returnDate:
          type: string
          description: ISO date string
        nights:
          type: integer
        price:
          type: number
          description: Price per person in rubles
        totalPrice:
          type: number
          description: Total price for all adults in rubles
        mealType:
          type: string
          description: Meal type (e.g. "Всё включено")
        imageUrl:
          type: string
        aiDescription:
          type: string
          description: AI-generated description of what to see and whether to book
        aiRecommendation:
          type: string
          description: Short AI verdict - стоит или нет
        bookingUrl:
          type: string
          description: URL to booking page on level.travel or partner site
      required:
        - id
        - destination
        - country
        - city
        - hotel
        - stars
        - departureDate
        - returnDate
        - nights
        - price
        - totalPrice
        - mealType
        - aiDescription
        - aiRecommendation
        - bookingUrl
    TourSearchResponse:
      type: object
      properties:
        tours:
          type: array
          items:
            $ref: "#/components/schemas/Tour"
        searchId:
          type: string
        departureCity:
          type: string
        budget:
          type: number
        totalFound:
          type: integer
      required:
        - tours
        - departureCity
        - budget
        - totalFound
    CitiesResponse:
      type: object
      properties:
        cities:
          type: array
          items:
            $ref: "#/components/schemas/City"
      required:
        - cities
    City:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        nameEn:
          type: string
      required:
        - id
        - name
        - nameEn
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'